requirements.txt
tinydb==4.7.1
SQLAlchemy==2.0.15
mysqlclient==2.1.1
requests==2.31.0
python-telegram-bot==20.3